Fix a race condition (and Replace InstanceKillProgressActor with Akka stream)
Summary:

The Akka Stream solution is arguably easier to reason about, and,
ensures that the event bus subscription is registered before any
kill operations are sent to Meosos.

The Akka Stream solution is arguably easier to reason about, and,
ensures that the event bus subscription is registered before any
kill operations are sent to Meosos.

Also-By: Karsten Jeschkies <jeskar@web.de>

Test Plan: scale up and down several times

Reviewers: jasongilanfarr, meichstedt

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D510